end studentDashboard feature

+++ Features/Diplomas/GetStudentDashboard/Dtos/DashboardStatsDto.cs
@@ -1,5 +1,5 @@
 ﻿namespace exam_system.Features.Diplomas.GetStudentDashboard.Dtos
 {
-    public record DashboardStatsDto(decimal AverageScore,decimal PassRate,int CompletedAttemptsCount, TimeSpan TotalTimeSpent);
+    public record DashboardStatsDto(double AverageScore,decimal PassRate,int CompletedAttemptsCount, TimeSpan TotalTimeSpent);
 }
